% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTCGAdata-package.R
\docType{package}
\name{prepTCGAdata-package}
\alias{prepTCGAdata}
\alias{prepTCGAdata-package}
\title{prepTCGAdata: Prepare TCGA Data for the SJV and SJVC Apps}
\description{
Downloads TCGA splice-junction data and clinical metadata via recount3 and
the GDC/cBioPortal APIs, annotates junctions against a GENCODE reference, and
prepares the metadata for overall-survival analysis.
}
\author{
\strong{Maintainer}: Phaedra Agius \email{phaedragius@gmail.com}

Authors:
\itemize{
  \item Phaedra Agius \email{phaedragius@gmail.com}
}

}
\keyword{internal}
